Update confidence_set_func.py
Added the option for Monte Carlo simulation so that the mean subtraction is always the true mean.
When doing bootstrap, the mean is forced to be the mean of the bootstrap sample instead of the full data prediction. This is due to that the mean prediction from full data might not be centered around the mean of the bootstrap.