Update confidence_set_func.py
Added in comparison method: naive method, this method uses the bootstrap samples at each point and evaluate whether it is below or above the threshold with greater than the lower bound. If it is greater than the lower bound, we declare it as inner set or outside of outer set.